+++ docs/changes.rst
@@ -6,6 +6,8 @@ Version 0.1.1
 
 To be released.
 
+- Added :class:`sider.types.Date` type.
+
 
 Version 0.1.0
 -------------